requirements.txt
alembic==1.6.5
click==8.0.1
flask>=2.0.3
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2>=3.0.3
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dateutil>=2.8.2
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.0.1
python-dotenv>=0.19.2
python-jose
authlib>=1.0
requests>=2.27.1
astral